Split `mendel-loader-server` into `mendel-loader` and `mendel-development-loader`
