+++ packages/mendel-development-middleware/package.json
@@ -18,7 +18,7 @@
     "fs-extra": "^0.30.0",
     "mendel-config": "1.0.0",
     "mendel-development": "1.0.0",
-    "mendel-loader-server": "1.0.0",
+    "mendel-development-loader": "1.0.0",
     "mendel-requirify": "1.0.0",
     "mendel-treenherit": "1.0.0",
     "stream-cache": "0.0.2",
