+++ packages/mendel-loader/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "mendel-loader",
+  "version": "1.0.0",
+  "description": "Mendel server side require hook.",
+  "main": "loader.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [
+    "mendel",
+    "browserify-plugin",
+    "build"
+  ],
+  "author": "David Gomez <david.garcia.mx@gmail.com>",
+  "license": "MIT",
+  "dependencies": {
+    "mendel-development": "1.0.0"
+  }
+}
